feat: Removed the instantly V1 support and replaced instantlyai into instantly
