+++ providers/instantly/delete_test.go
@@ -16,59 +16,60 @@ import (
 func TestDelete(t *testing.T) { // nolint:funlen,cyclop
 	t.Parallel()
 
-	responseNotFoundErr := testutils.DataFromFile(t, "delete-tag-missing.json")
-	responseTag := testutils.DataFromFile(t, "delete-tag.json")
+	errorNotFound := testutils.DataFromFile(t, "delete-missing-custom-tags.json")
 
 	tests := []testroutines.Delete{
 		{
-			Name:         "Write object must be included",
+			Name:         "Delete object must be included",
 			Server:       mockserver.Dummy(),
 			ExpectedErrs: []error{common.ErrMissingObjects},
 		},
 		{
 			Name:         "Write object and its ID must be included",
-			Input:        common.DeleteParams{ObjectName: "tags"},
+			Input:        common.DeleteParams{ObjectName: "custom-tags"},
 			Server:       mockserver.Dummy(),
 			ExpectedErrs: []error{common.ErrMissingRecordID},
 		},
 		{
-			Name:   "Cannot remove unknown object",
-			Input:  common.DeleteParams{ObjectName: "coupons", RecordId: "132"},
+			Name:   "Object name is not supported",
+			Input:  common.DeleteParams{ObjectName: "background-jobs", RecordId: "675266e304a8e55b17f0228b"},
 			Server: mockserver.Dummy(),
 			ExpectedErrs: []error{
 				common.ErrOperationNotSupportedForObject,
 			},
 		},
 		{
-			Name:  "Cannot remove missing tag",
-			Input: common.DeleteParams{ObjectName: "tags", RecordId: "5043"},
-			Server: mockserver.Fixed{
-				Setup:  mockserver.ContentJSON(),
-				Always: mockserver.Response(http.StatusNotFound, responseNotFoundErr),
-			}.Server(),
-			ExpectedErrs: []error{
-				common.ErrBadRequest,
-				errors.New(`Not Found`), // nolint:goerr113
-			},
-		},
-		{
 			Name:  "Successful delete",
-			Input: common.DeleteParams{ObjectName: "tags", RecordId: "5043"},
+			Input: common.DeleteParams{ObjectName: "custom-tags", RecordId: "8cba2d22-cdd0-453f-ac69-4c43f71942e5"},
 			Server: mockserver.Conditional{
 				Setup: mockserver.ContentJSON(),
 				If: mockcond.And{
+					mockcond.Path("/api/v2/custom-tags/8cba2d22-cdd0-453f-ac69-4c43f71942e5"),
 					mockcond.MethodDELETE(),
-					mockcond.Path("/api/v1/custom-tag/5043"),
+					mockcond.Body("null"),
 				},
-				Then: mockserver.Response(http.StatusOK, responseTag),
+				Then: mockserver.Response(http.StatusOK),
 			}.Server(),
 			Expected:     &common.DeleteResult{Success: true},
 			ExpectedErrs: nil,
 		},
+		{
+			Name:  "Error on deleting missing record",
+			Input: common.DeleteParams{ObjectName: "custom-tags", RecordId: "4fa0ca9d-f205-4a68-8756-5ad62123a53a"},
+			Server: mockserver.Fixed{
+				Setup:  mockserver.ContentJSON(),
+				Always: mockserver.Response(http.StatusNotFound, errorNotFound),
+			}.Server(),
+			ExpectedErrs: []error{
+				common.ErrBadRequest,
+				errors.New( // nolint:goerr113
+					" Not Found",
+				),
+			},
+		},
 	}
 
 	for _, tt := range tests {
-		// nolint:varnamelen
 		t.Run(tt.Name, func(t *testing.T) {
 			t.Parallel()
