+++ providers/instantly/handlers.go
@@ -1,4 +1,4 @@
-package instantlyai
+package instantly
 
 import (
 	"bytes"
