--- providers/instantly/metadata.go
@@ -1,14 +1,0 @@
-package instantly
-
-import (
-	"context"
-
-	"github.com/amp-labs/connectors/common"
-	"github.com/amp-labs/connectors/providers/instantly/metadata"
-)
-
-func (c *Connector) ListObjectMetadata(
-	ctx context.Context, objectNames []string,
-) (*common.ListObjectMetadataResult, error) {
-	return metadata.Schemas.Select(c.Module.ID, objectNames)
-}
