+++ providers/instantly/metadata_test.go
@@ -1,16 +1,24 @@
 package instantly
 
 import (
+	"net/http"
 	"testing"
 
 	"github.com/amp-labs/connectors"
 	"github.com/amp-labs/connectors/common"
+	"github.com/amp-labs/connectors/test/utils/mockutils"
+	"github.com/amp-labs/connectors/test/utils/mockutils/mockcond"
 	"github.com/amp-labs/connectors/test/utils/mockutils/mockserver"
 	"github.com/amp-labs/connectors/test/utils/testroutines"
+	"github.com/amp-labs/connectors/test/utils/testutils"
 )
 
-func TestListObjectMetadata(t *testing.T) { // nolint:funlen,gocognit,cyclop
+func TestListObjectMetadata(t *testing.T) { // nolint:funlen,gocognit,cyclop,maintidx
 	t.Parallel()
+
+	backgroundJobsResponse := testutils.DataFromFile(t, "background-jobs.json")
+	customTagsResponse := testutils.DataFromFile(t, "custom-tags.json")
+	leadListsResponse := testutils.DataFromFile(t, "lead-lists.json")
 
 	tests := []testroutines.Metadata{
 		{
@@ -20,39 +28,219 @@ func TestListObjectMetadata(t *testing.T) { // nolint:funlen,gocognit,cyclop
 			ExpectedErrs: []error{common.ErrMissingObjects},
 		},
 		{
-			Name:       "Unknown object requested",
-			Input:      []string{"butterflies"},
-			Server:     mockserver.Dummy(),
-			Comparator: testroutines.ComparatorSubsetMetadata,
-			Expected: &common.ListObjectMetadataResult{
-				Errors: map[string]error{
-					"butterflies": common.ErrObjectNotSupported,
-				},
-			},
-		},
-		{
-			Name:       "Successfully describe multiple objects with metadata",
-			Input:      []string{"campaigns", "emails"},
-			Server:     mockserver.Dummy(),
+			Name:  "Successfully describe multiple objects with metadata",
+			Input: []string{"background-jobs", "custom-tags", "lead-lists"},
+			Server: mockserver.Switch{
+				Setup: mockserver.ContentJSON(),
+				Cases: []mockserver.Case{{
+					If:   mockcond.Path("v2/background-jobs"),
+					Then: mockserver.Response(http.StatusOK, backgroundJobsResponse),
+				}, {
+					If:   mockcond.Path("v2/custom-tags"),
+					Then: mockserver.Response(http.StatusOK, customTagsResponse),
+				}, {
+					If:   mockcond.Path("v2/lead-lists"),
+					Then: mockserver.Response(http.StatusOK, leadListsResponse),
+				}},
+			}.Server(),
 			Comparator: testroutines.ComparatorSubsetMetadata,
 			Expected: &common.ListObjectMetadataResult{
 				Result: map[string]common.ObjectMetadata{
-					"emails": {
-						DisplayName: "Emails",
+					"background-jobs": {
+						DisplayName: "Background-jobs",
+						Fields: map[string]common.FieldMetadata{
+							"id": {
+								DisplayName:  "id",
+								ValueType:    "other",
+								ProviderType: "",
+								ReadOnly:     false,
+								Values:       nil,
+							},
+							"workspace_id": {
+								DisplayName:  "workspace_id",
+								ValueType:    "other",
+								ProviderType: "",
+								ReadOnly:     false,
+								Values:       nil,
+							},
+							"type": {
+								DisplayName:  "type",
+								ValueType:    "other",
+								ProviderType: "",
+								ReadOnly:     false,
+								Values:       nil,
+							},
+							"entity_id": {
+								DisplayName:  "entity_id",
+								ValueType:    "other",
+								ProviderType: "",
+								ReadOnly:     false,
+								Values:       nil,
+							},
+							"entity_type": {
+								DisplayName:  "entity_type",
+								ValueType:    "other",
+								ProviderType: "",
+								ReadOnly:     false,
+								Values:       nil,
+							},
+							"data": {
+								DisplayName:  "data",
+								ValueType:    "other",
+								ProviderType: "",
+								ReadOnly:     false,
+								Values:       nil,
+							},
+							"progress": {
+								DisplayName:  "progress",
+								ValueType:    "other",
+								ProviderType: "",
+								ReadOnly:     false,
+								Values:       nil,
+							},
+							"status": {
+								DisplayName:  "status",
+								ValueType:    "other",
+								ProviderType: "",
+								ReadOnly:     false,
+								Values:       nil,
+							},
+							"created_at": {
+								DisplayName:  "created_at",
+								ValueType:    "other",
+								ProviderType: "",
+								ReadOnly:     false,
+								Values:       nil,
+							},
+							"updated_at": {
+								DisplayName:  "updated_at",
+								ValueType:    "other",
+								ProviderType: "",
+								ReadOnly:     false,
+								Values:       nil,
+							},
+						},
 						FieldsMap: map[string]string{
-							"is_unread":             "Is Unread",
-							"ue_type":               "Email Type",
-							"message_id":            "Message ID",
-							"campaign_id":           "Campaign ID",
-							"from_address_email":    "Sender's Email Address",
-							"to_address_email_list": "Receivers Email Address",
+							"id":           "id",
+							"workspace_id": "workspace_id",
+							"type":         "type",
+							"entity_id":    "entity_id",
+							"entity_type":  "entity_type",
+							"data":         "data",
+							"progress":     "progress",
+							"status":       "status",
+							"created_at":   "created_at",
+							"updated_at":   "updated_at",
 						},
 					},
-					"campaigns": {
-						DisplayName: "Campaigns",
+					"custom-tags": {
+						DisplayName: "Custom-tags",
+						Fields: map[string]common.FieldMetadata{
+							"id": {
+								DisplayName:  "id",
+								ValueType:    "other",
+								ProviderType: "",
+								ReadOnly:     false,
+								Values:       nil,
+							},
+							"timestamp_created": {
+								DisplayName:  "timestamp_created",
+								ValueType:    "other",
+								ProviderType: "",
+								ReadOnly:     false,
+								Values:       nil,
+							},
+							"timestamp_updated": {
+								DisplayName:  "timestamp_updated",
+								ValueType:    "other",
+								ProviderType: "",
+								ReadOnly:     false,
+								Values:       nil,
+							},
+							"organization_id": {
+								DisplayName:  "organization_id",
+								ValueType:    "other",
+								ProviderType: "",
+								ReadOnly:     false,
+								Values:       nil,
+							},
+							"label": {
+								DisplayName:  "label",
+								ValueType:    "other",
+								ProviderType: "",
+								ReadOnly:     false,
+								Values:       nil,
+							},
+							"description": {
+								DisplayName:  "description",
+								ValueType:    "other",
+								ProviderType: "",
+								ReadOnly:     false,
+								Values:       nil,
+							},
+						},
 						FieldsMap: map[string]string{
-							"id":   "ID",
-							"name": "Name",
+							"id":                "id",
+							"timestamp_created": "timestamp_created",
+							"timestamp_updated": "timestamp_updated",
+							"organization_id":   "organization_id",
+							"label":             "label",
+							"description":       "description",
+						},
+					},
+					"lead-lists": {
+						DisplayName: "Lead-lists",
+						Fields: map[string]common.FieldMetadata{
+							"id": {
+								DisplayName:  "id",
+								ValueType:    "other",
+								ProviderType: "",
+								ReadOnly:     false,
+								Values:       nil,
+							},
+							"organization_id": {
+								DisplayName:  "organization_id",
+								ValueType:    "other",
+								ProviderType: "",
+								ReadOnly:     false,
+								Values:       nil,
+							},
+							"has_enrichment_task": {
+								DisplayName:  "has_enrichment_task",
+								ValueType:    "other",
+								ProviderType: "",
+								ReadOnly:     false,
+								Values:       nil,
+							},
+							"owned_by": {
+								DisplayName:  "owned_by",
+								ValueType:    "other",
+								ProviderType: "",
+								ReadOnly:     false,
+								Values:       nil,
+							},
+							"name": {
+								DisplayName:  "name",
+								ValueType:    "other",
+								ProviderType: "",
+								ReadOnly:     false,
+								Values:       nil,
+							},
+							"timestamp_created": {
+								DisplayName:  "timestamp_created",
+								ValueType:    "other",
+								ProviderType: "",
+								ReadOnly:     false,
+								Values:       nil,
+							},
+						},
+						FieldsMap: map[string]string{
+							"id":                  "id",
+							"organization_id":     "organization_id",
+							"has_enrichment_task": "has_enrichment_task",
+							"owned_by":            "owned_by",
+							"name":                "name",
+							"timestamp_created":   "timestamp_created",
 						},
 					},
 				},
@@ -63,7 +251,6 @@ func TestListObjectMetadata(t *testing.T) { // nolint:funlen,gocognit,cyclop
 	}
 
 	for _, tt := range tests {
-		// nolint:varnamelen
 		t.Run(tt.Name, func(t *testing.T) {
 			t.Parallel()
 
@@ -72,4 +259,19 @@ func TestListObjectMetadata(t *testing.T) { // nolint:funlen,gocognit,cyclop
 			})
 		})
 	}
+}
+
+func constructTestConnector(serverURL string) (*Connector, error) {
+	connector, err := NewConnector(common.ConnectorParams{
+		Module:              common.ModuleRoot,
+		AuthenticatedClient: mockutils.NewClient(),
+	})
+	if err != nil {
+		return nil, err
+	}
+
+	// for testing we want to redirect calls to our mock server
+	connector.SetBaseURL(mockutils.ReplaceURLOrigin(connector.HTTPClient().Base, serverURL))
+
+	return connector, nil
 }
