--- providers/instantly/params.go
@@ -1,56 +1,0 @@
-package instantly
-
-import (
-	"context"
-	"errors"
-	"net/http"
-
-	"github.com/amp-labs/connectors/common"
-	"github.com/amp-labs/connectors/common/paramsbuilder"
-	"github.com/amp-labs/connectors/providers"
-)
-
-const (
-	// DefaultPageSize is number of elements per page.
-	DefaultPageSize = 100
-)
-
-// Option is a function which mutates the connector configuration.
-type Option = func(params *parameters)
-
-// parameters surface options by delegation.
-type parameters struct {
-	paramsbuilder.Client
-}
-
-func newParams(opts []Option) (*common.ConnectorParams, error) { // nolint:unused
-	oldParams, err := paramsbuilder.Apply(parameters{}, opts)
-	if err != nil {
-		return nil, err
-	}
-
-	return &common.ConnectorParams{
-		AuthenticatedClient: oldParams.Client.Caller.Client,
-	}, nil
-}
-
-func (p parameters) ValidateParams() error {
-	return errors.Join(
-		p.Client.ValidateParams(),
-	)
-}
-
-func WithClient(ctx context.Context, client *http.Client,
-	apiKey string,
-	opts ...common.QueryParamAuthClientOption,
-) Option {
-	return func(params *parameters) {
-		params.WithApiKeyQueryParamClient(ctx, client, providers.Instantly, apiKey, opts...)
-	}
-}
-
-func WithAuthenticatedClient(client common.AuthenticatedHTTPClient) Option {
-	return func(params *parameters) {
-		params.WithAuthenticatedClient(client)
-	}
-}
