--- providers/instantly/test/delete-tag-missing.json
@@ -1,6 +1,0 @@
-{
-    "statusCode": 404,
-    "error": "Not Found",
-    "message": "Tag not found"
-}
-
