--- providers/instantly/test/delete-tag.json
@@ -1,9 +1,0 @@
-{
-    "id": "bc5e72fc-95dc-4bfb-979f-bfe897dfc0eb",
-    "timestamp_created": "2024-08-14T17:41:57.998Z",
-    "timestamp_updated": "2024-08-14T17:41:57.998Z",
-    "organization_id": "803a064c-a636-49fe-bc45-5043da7a4ee7",
-    "label": "High Delivery",
-    "description": "High Delivery Accounts"
-}
-
