--- providers/instantly/test/invalid-path.json
@@ -1,6 +1,0 @@
-{
-    "message": "Route GET:/api/butterfly?api_key=<SENSITIVE_DATA> not found",
-    "error": "Not Found",
-    "statusCode": 404
-}
-
