--- providers/instantly/test/read-campaigns.json
@@ -1,11 +1,0 @@
-[
-    {
-        "id": "27dd47ff-5a78-4377-a1eb-98f593f37219",
-        "name": "Second Campaign"
-    },
-    {
-        "id": "65d890fe-ae4d-43d0-b014-af0e89b6281f",
-        "name": "My Campaign"
-    }
-]
-
