--- providers/instantly/test/write-blocklist-entry.json
@@ -1,6 +1,0 @@
-{
-  "status": "success",
-  "entries_added": 2,
-  "already_in_blocklist": 0,
-  "blocklist_id": "cf8fd143-08c3-438e-a396-491aa1ced9d4"
-}
