--- providers/instantly/test/write-lead-bad-request.json
@@ -1,3 +1,0 @@
-{
-  "error": "Leads array is empty"
-}
