--- providers/instantly/test/write-tag.json
@@ -1,9 +1,0 @@
-{
-    "id": "f6825fcf-c51b-4724-937b-0814ed02af83",
-    "timestamp_created": "2024-08-14T17:40:35.627Z",
-    "timestamp_updated": "2024-08-14T17:40:35.627Z",
-    "organization_id": "803a064c-a636-49fe-bc45-5043da7a4ee7",
-    "label": "High Delivery",
-    "description": "High Delivery Accounts"
-}
-
