--- providers/instantly/test/write-unibox-reply.json
@@ -1,4 +1,0 @@
-{
-  "status": "success",
-  "message_id": "19e9d7f9-bd4f-45aa-8d77-9eecc9bd3f9a"
-}
