+++ providers/instantly/write_test.go
@@ -1,7 +1,6 @@
 package instantly
 
 import (
-	"errors"
 	"net/http"
 	"testing"
 
@@ -13,18 +12,11 @@ import (
 	"github.com/amp-labs/connectors/test/utils/testutils"
 )
 
-func TestWrite(t *testing.T) { // nolint:funlen,cyclop
+func TestWrite(t *testing.T) { // nolint:funlen,gocognit,cyclop,maintidx
 	t.Parallel()
 
-	responseBlocklistEntry := testutils.DataFromFile(t, "write-blocklist-entry.json")
-
-	responseReply := testutils.DataFromFile(t, "write-unibox-reply.json")
-
-	responseLeadErr := testutils.DataFromFile(t, "write-lead-bad-request.json")
-	responseLead := testutils.DataFromFile(t, "write-lead.json")
-
-	responseTagErr := testutils.DataFromFile(t, "write-tag-bad-request.json")
-	responseTag := testutils.DataFromFile(t, "write-tag.json")
+	customTagsResponse := testutils.DataFromFile(t, "write_custom_tags.json")
+	accountsResponse := testutils.DataFromFile(t, "write_accounts.json")
 
 	tests := []testroutines.Write{
 		{
@@ -33,129 +25,89 @@ func TestWrite(t *testing.T) { // nolint:funlen,cyclop
 			ExpectedErrs: []error{common.ErrMissingObjects},
 		},
 		{
-			Name:         "Write needs data payload",
-			Input:        common.WriteParams{ObjectName: "notes"},
-			Server:       mockserver.Dummy(),
-			ExpectedErrs: []error{common.ErrMissingRecordData},
-		},
-		{
-			Name:     "Unknown object name is not supported",
-			Input:    common.WriteParams{ObjectName: "orders", RecordData: "dummy"},
-			Server:   mockserver.Dummy(),
-			Expected: nil,
-			ExpectedErrs: []error{
-				common.ErrOperationNotSupportedForObject,
-			},
-		},
-		{
-			Name:  "Create Blocklist Entry",
-			Input: common.WriteParams{ObjectName: "blocklist-entries", RecordData: "dummy"},
+			Name:  "creating the custom tags",
+			Input: common.WriteParams{ObjectName: "custom-tags", RecordData: "dummy"},
 			Server: mockserver.Conditional{
 				Setup: mockserver.ContentJSON(),
-				If:    mockcond.MethodPOST(),
-				Then:  mockserver.Response(http.StatusOK, responseBlocklistEntry),
+				If: mockcond.And{
+					mockcond.Path("/api/v2/custom-tags"),
+					mockcond.MethodPOST(),
+				},
+				Then: mockserver.Response(http.StatusOK, customTagsResponse),
 			}.Server(),
 			Expected: &common.WriteResult{
 				Success:  true,
-				RecordId: "cf8fd143-08c3-438e-a396-491aa1ced9d4",
+				RecordId: "0196837f-d659-7496-854c-84187bb6f708",
 				Errors:   nil,
-				Data:     nil,
+				Data: map[string]any{
+					"id":                "0196837f-d659-7496-854c-84187bb6f708",
+					"timestamp_created": "2025-04-29T21:41:55.417Z",
+					"timestamp_updated": "2025-04-29T21:41:55.417Z",
+					"organization_id":   "0196837f-d659-7496-854c-841903f3a321",
+					"label":             "Important",
+					"description":       "Used for marking important items",
+				},
 			},
 			ExpectedErrs: nil,
 		},
 		{
-			Name:  "Create Unibox Reply",
-			Input: common.WriteParams{ObjectName: "unibox-replies", RecordData: "dummy"},
+			Name: "Update custom tags as PATCH",
+			Input: common.WriteParams{
+				ObjectName: "custom-tags",
+				RecordId:   "0196837f-d659-7496-854c-84187bb6f708",
+				RecordData: "dummy",
+			},
 			Server: mockserver.Conditional{
 				Setup: mockserver.ContentJSON(),
-				If:    mockcond.MethodPOST(),
-				Then:  mockserver.Response(http.StatusOK, responseReply),
+				If: mockcond.And{
+					mockcond.Path("/api/v2/custom-tags/0196837f-d659-7496-854c-84187bb6f708"),
+					mockcond.MethodPATCH(),
+				},
+				Then: mockserver.Response(http.StatusOK, customTagsResponse),
 			}.Server(),
 			Expected: &common.WriteResult{
 				Success:  true,
-				RecordId: "19e9d7f9-bd4f-45aa-8d77-9eecc9bd3f9a",
+				RecordId: "0196837f-d659-7496-854c-84187bb6f708",
 				Errors:   nil,
-				Data:     nil,
-			},
-			ExpectedErrs: nil,
-		},
-
-		{
-			Name:  "Invalid Lead creation",
-			Input: common.WriteParams{ObjectName: "leads", RecordData: "dummy"},
-			Server: mockserver.Fixed{
-				Setup:  mockserver.ContentJSON(),
-				Always: mockserver.Response(http.StatusNotFound, responseLeadErr),
-			}.Server(),
-			ExpectedErrs: []error{
-				common.ErrBadRequest,
-				errors.New("Leads array is empty"), // nolint:goerr113
-			},
-		},
-		{
-			Name:  "Create Lead",
-			Input: common.WriteParams{ObjectName: "leads", RecordData: "dummy"},
-			Server: mockserver.Conditional{
-				Setup: mockserver.ContentJSON(),
-				If:    mockcond.MethodPOST(),
-				Then:  mockserver.Response(http.StatusOK, responseLead),
-			}.Server(),
-			Expected: &common.WriteResult{
-				Success:  true,
-				RecordId: "", // lead doesn't have ID
-				Errors:   nil,
-				Data:     nil,
+				Data: map[string]any{
+					"id":                "0196837f-d659-7496-854c-84187bb6f708",
+					"timestamp_created": "2025-04-29T21:41:55.417Z",
+					"timestamp_updated": "2025-04-29T21:41:55.417Z",
+					"organization_id":   "0196837f-d659-7496-854c-841903f3a321",
+					"label":             "Important",
+					"description":       "Used for marking important items",
+				},
 			},
 			ExpectedErrs: nil,
 		},
 		{
-			Name:  "Invalid Tag creation",
-			Input: common.WriteParams{ObjectName: "tags", RecordData: "dummy"},
-			Server: mockserver.Fixed{
-				Setup:  mockserver.ContentJSON(),
-				Always: mockserver.Response(http.StatusNotFound, responseTagErr),
-			}.Server(),
-			ExpectedErrs: []error{
-				common.ErrBadRequest,
-				errors.New("Bad Request"), // nolint:goerr113
-			},
-		},
-		{
-			Name:  "Create Tag acts as POST",
-			Input: common.WriteParams{ObjectName: "tags", RecordData: "dummy"},
+			Name:  "Creating the accounts",
+			Input: common.WriteParams{ObjectName: "accounts", RecordData: "dummy"},
 			Server: mockserver.Conditional{
 				Setup: mockserver.ContentJSON(),
-				If:    mockcond.MethodPOST(),
-				Then:  mockserver.Response(http.StatusOK, responseTag),
+				If: mockcond.And{
+					mockcond.Path("/api/v2/accounts"),
+					mockcond.MethodPOST(),
+				},
+				Then: mockserver.Response(http.StatusOK, accountsResponse),
 			}.Server(),
 			Expected: &common.WriteResult{
 				Success:  true,
-				RecordId: "f6825fcf-c51b-4724-937b-0814ed02af83",
+				RecordId: "",
 				Errors:   nil,
-				Data:     nil,
-			},
-			ExpectedErrs: nil,
-		},
-		{
-			Name:  "Update tag acts as PATCH",
-			Input: common.WriteParams{ObjectName: "tags", RecordId: "885633", RecordData: "dummy"},
-			Server: mockserver.Conditional{
-				Setup: mockserver.ContentJSON(),
-				If:    mockcond.MethodPATCH(),
-				Then:  mockserver.Response(http.StatusOK, responseTag),
-			}.Server(),
-			Expected: &common.WriteResult{
-				Success:  true,
-				RecordId: "f6825fcf-c51b-4724-937b-0814ed02af83",
-				Errors:   nil,
-				Data:     nil,
+				Data: map[string]any{
+					"email":             "user@example.com",
+					"timestamp_created": "2025-04-29T21:41:55.421Z",
+					"timestamp_updated": "2025-04-29T21:41:55.421Z",
+					"first_name":        "John",
+					"last_name":         "Doe",
+				},
 			},
 			ExpectedErrs: nil,
 		},
 	}
 
 	for _, tt := range tests {
-		// nolint:varnamelen
 		t.Run(tt.Name, func(t *testing.T) {
 			t.Parallel()
