--- providers/instantlyai/connector.go
@@ -1,90 +1,0 @@
-package instantlyai
-
-import (
-	_ "embed"
-
-	"github.com/amp-labs/connectors/common"
-	"github.com/amp-labs/connectors/common/interpreter"
-	"github.com/amp-labs/connectors/internal/components"
-	"github.com/amp-labs/connectors/internal/components/deleter"
-	"github.com/amp-labs/connectors/internal/components/operations"
-	"github.com/amp-labs/connectors/internal/components/reader"
-	"github.com/amp-labs/connectors/internal/components/schema"
-	"github.com/amp-labs/connectors/internal/components/writer"
-	"github.com/amp-labs/connectors/providers"
-	"github.com/amp-labs/connectors/providers/instantlyai/metadata"
-)
-
-type Connector struct {
-	// Basic connector
-	*components.Connector
-
-	// Require authenticated client
-	common.RequireAuthenticatedClient
-
-	// Supported operations
-	components.SchemaProvider
-	components.Reader
-	components.Writer
-	components.Deleter
-}
-
-const apiVersion = "v2"
-
-func NewConnector(params common.ConnectorParams) (*Connector, error) {
-	// Create base connector with provider info
-	return components.Initialize(providers.InstantlyAI, params, constructor)
-}
-
-func constructor(base *components.Connector) (*Connector, error) {
-	connector := &Connector{Connector: base}
-
-	// Set the metadata provider for the connector
-	connector.SchemaProvider = schema.NewOpenAPISchemaProvider(connector.ProviderContext.Module(), metadata.Schemas)
-
-	registry, err := components.NewEndpointRegistry(supportedOperations())
-	if err != nil {
-		return nil, err
-	}
-
-	connector.Reader = reader.NewHTTPReader(
-		connector.HTTPClient().Client,
-		registry,
-		connector.ProviderContext.Module(),
-		operations.ReadHandlers{
-			BuildRequest:  connector.buildReadRequest,
-			ParseResponse: connector.parseReadResponse,
-			ErrorHandler: interpreter.ErrorHandler{
-				JSON: interpreter.NewFaultyResponder(errorFormats, nil),
-			}.Handle,
-		},
-	)
-
-	connector.Writer = writer.NewHTTPWriter(
-		connector.HTTPClient().Client,
-		registry,
-		connector.ProviderContext.Module(),
-		operations.WriteHandlers{
-			BuildRequest:  connector.buildWriteRequest,
-			ParseResponse: connector.parseWriteResponse,
-			ErrorHandler: interpreter.ErrorHandler{
-				JSON: interpreter.NewFaultyResponder(errorFormats, nil),
-			}.Handle,
-		},
-	)
-
-	connector.Deleter = deleter.NewHTTPDeleter(
-		connector.HTTPClient().Client,
-		registry,
-		connector.ProviderContext.Module(),
-		operations.DeleteHandlers{
-			BuildRequest:  connector.buildDeleteRequest,
-			ParseResponse: connector.parseDeleteResponse,
-			ErrorHandler: interpreter.ErrorHandler{
-				JSON: interpreter.NewFaultyResponder(errorFormats, nil),
-			}.Handle,
-		},
-	)
-
-	return connector, nil
-}
