--- providers/instantlyai/errors.go
@@ -1,29 +1,0 @@
-package instantlyai
-
-import (
-	"fmt"
-
-	"github.com/amp-labs/connectors/common/interpreter"
-)
-
-var errorFormats = interpreter.NewFormatSwitch( // nolint:gochecknoglobals
-	[]interpreter.FormatTemplate{
-		{
-			MustKeys: nil,
-			Template: func() interpreter.ErrorDescriptor { return &ResponseError{} },
-		},
-	}...,
-)
-
-type ResponseError struct {
-	// Message usually has API Key query parameter. This shouldn't be returned.
-	Message string `json:"message"`
-	// Error is very general.
-	Error      string `json:"error"`
-	StatusCode int    `json:"statusCode"`
-}
-
-func (r ResponseError) CombineErr(base error) error {
-	// Error field is the safest to return, though not very useful.
-	return fmt.Errorf("%w: %v", base, r.Error)
-}
