+++ test/instantly/connector.go
@@ -2,8 +2,8 @@ package instantly
 
 import (
 	"context"
-	"net/http"
 
+	"github.com/amp-labs/connectors/common"
 	"github.com/amp-labs/connectors/common/scanning/credscanning"
 	"github.com/amp-labs/connectors/providers"
 	"github.com/amp-labs/connectors/providers/instantly"
@@ -11,16 +11,16 @@ import (
 )
 
 func GetInstantlyConnector(ctx context.Context) *instantly.Connector {
-	filePath := credscanning.LoadPath(providers.Instantly)
+	filePath := credscanning.LoadPath(providers.InstantlyAI)
 	reader := utils.MustCreateProvCredJSON(filePath, false)
 
-	conn, err := instantly.NewConnector(
-		instantly.WithClient(ctx, http.DefaultClient,
-			reader.Get(credscanning.Fields.ApiKey),
-		),
-	)
+	client := utils.NewAPIKeyClient(ctx, reader, providers.InstantlyAI)
+
+	conn, err := instantly.NewConnector(common.ConnectorParams{
+		AuthenticatedClient: client,
+	})
 	if err != nil {
-		utils.Fail("error creating connector", "error", err)
+		utils.Fail("error creating Instantly App connector", "error", err)
 	}
 
 	return conn
