+++ test/instantly/delete/delete.go
@@ -8,8 +8,8 @@ import (
 	"os"
 
 	"github.com/amp-labs/connectors/common"
-	ap "github.com/amp-labs/connectors/providers/instantlyai"
-	"github.com/amp-labs/connectors/test/instantlyai"
+	ap "github.com/amp-labs/connectors/providers/instantly"
+	"github.com/amp-labs/connectors/test/instantly"
 )
 
 func main() {
@@ -33,7 +33,7 @@ func MainFn() int {
 }
 
 func testCustomTags(ctx context.Context) error {
-	conn := instantlyai.GetInstantlyAIConnector(ctx)
+	conn := instantly.GetInstantlyConnector(ctx)
 
 	slog.Info("Deleting the custom tags")
 
@@ -57,7 +57,7 @@ func testCustomTags(ctx context.Context) error {
 }
 
 func testLeadLists(ctx context.Context) error {
-	conn := instantlyai.GetInstantlyAIConnector(ctx)
+	conn := instantly.GetInstantlyConnector(ctx)
 
 	slog.Info("Deleting the lead lists")
